Anderungen auf vt001175

+++ test/flate_plate/flate_plate_extract.F90
@@ -12,8 +12,8 @@ integer :: i_start,i_end
 integer :: j_start,j_end
 
 logical :: fexists
-integer(kind=CGSIZE_T) :: cgns_file,ierror,cgns_base,PhysDim,cgns_zone,zonetype
-integer(kind=CGSIZE_T) :: nSol,nVar_in,datatype,data_location,var,sol
+integer :: cgns_file,ierror,cgns_base,PhysDim,cgns_zone,zonetype
+integer :: nSol,nVar_in,datatype,data_location,var,sol
 integer(kind=CGSIZE_T),allocatable :: isize(:,:),istart(:)
 character(len=32)  :: solname,cgns_git_basename
 character(len=*), parameter :: file_data_in = "data_out.cgns"
